// packages/nx-jest/src/executors/jest/schema.d.ts
export interface NxJestExecutorOptions {
    codeCoverage?: boolean;
    config?: string;
    detectOpenHandles?: boolean;
    jestConfig: string;
    testFile?: string;
    setupFile?: string;
    bail?: boolean | number;
    ci?: boolean;
    color?: boolean;
    clearCache?: boolean;
    findRelatedTests?: string;
    json?: boolean;
    maxWorkers?: number | string;
    onlyChanged?: boolean;
    changedSince?: string;
    outputFile?: string;
    passWithNoTests?: boolean;
    runInBand?: boolean;
    showConfig?: boolean;
    silent?: boolean;
    testNamePattern?: string;
    testPathIgnorePatterns?: string[];
    testPathPattern?: string[];
    colors?: boolean;
    reporters?: string[];
    verbose?: boolean;
    coverageReporters?: string[];
    coverageDirectory?: string;
    testResultsProcessor?: string;
    updateSnapshot?: boolean;
    useStderr?: boolean;
    watch?: boolean;
    watchAll?: boolean;
    testLocationInResults?: boolean;
    testTimeout?: number;
    testFromSource?: boolean;
}